refactor(utils/misc): move functions and adjust usages also renamed some functions for clarity for #5187
