+++ frontend/src/ts/test/replay.ts
@@ -1,6 +1,7 @@
 import config from "../config";
 import * as Sound from "../controllers/sound-controller";
 import * as TestInput from "./test-input";
+import * as Misc from "../utils/misc";
 
 type ReplayAction =
   | "correctLetter"
@@ -270,7 +271,7 @@ function playReplay(): void {
 
   let swTime = Math.round(lastTime / 1000); //starting time
   const swEndTime = Math.round(
-    (replayData[replayData.length - 1] as Replay).time / 1000
+    (Misc.lastElementFromArray(replayData) as Replay).time / 1000
   );
   while (swTime <= swEndTime) {
     const time = swTime;
@@ -299,7 +300,7 @@ function playReplay(): void {
         "aria-label",
         "Start replay"
       );
-    }, (replayData[replayData.length - 1] as Replay).time - lastTime)
+    }, (Misc.lastElementFromArray(replayData) as Replay).time - lastTime)
   );
 }
